--- app/src/main/java/com/example/culturegram/Heritage.kt
@@ -1,9 +1,0 @@
-package com.example.culturegram
-
-class SakeBrewery (
-    val name: String,
-    val latitude: Double,
-    val longitude: Double,
-    val distance: Double,
-    val yet: Int,
-)
