+++ app/src/main/java/com/example/culturegram/SakeBrewery.kt
@@ -1,0 +1,20 @@
+package com.example.culturegram
+
+class SakeBrewery (
+    val name: String,
+    val latitude: Double,
+    val longitude: Double,
+    val distance: Double,
+    val yet: Int,
+)
+
+class SakeBrewery2 (
+    val ID: Int,
+    val breweryName: String,
+    val sakeName: String,
+    val latitude: Double,
+    val longitude: Double,
+    val distance: Double,
+    val yet: Int,
+    val attr: String
+)
